统一equal_op
